Change sizes to Long and fix a few doc comments

+++ src/main/scala/scodec/codecs/FixedSizeCodec.scala
@@ -6,7 +6,7 @@ import scalaz.syntax.std.either._
 
 import scodec.bits.BitVector
 
-private[codecs] final class FixedSizeCodec[A](size: Int, codec: Codec[A]) extends Codec[A] {
+private[codecs] final class FixedSizeCodec[A](size: Long, codec: Codec[A]) extends Codec[A] {
 
   override def encode(a: A) = for {
     encoded <- codec.encode(a)
